Test images
There are 3 versions of each image: one with a black background, one with a white background, and one with a transparent background. The the variations are otherwise identical.

The images follow a naming scheme to make them easy to reference in code. Each meteorite has been assigned an arbitrary number, followed by a single character denoting the background type: 'b' for black, 'w' for white, and 'c' for clear (or transparent).